Organize translation script
