+++ requirements.txt
@@ -7,6 +7,7 @@ openai
 nlpaug
 nltk
 num2words
+text2num
 tqdm
 soundfile
 librosa==0.9.2
