+++ run_txt_aug.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import argparse
 import random
-from text.aug import create_augmentation_sequence
+from text.aug import create_augmentation_sequence, translate_sentences_api
 from utils.text import (
     post_process_sentences,
     print_sentences_comparison,
@@ -52,6 +52,12 @@ if __name__ == "__main__":
         help="Random seed (default: 451)",
     )
     parser.add_argument(
+        "--slang",
+        type=str,
+        default="pt",
+        help="Source language for translation (default: pt)",
+    )
+    parser.add_argument(
         "--lang",
         type=str,
         default="en",
@@ -83,6 +89,10 @@ if __name__ == "__main__":
 
     sentences = read_sentences_corpus(args.corpus, max_sentences=args.maxs)
     print(f"Read {len(sentences)} sentences from {args.corpus}")
+
+    if args.aug == ["translate"]:
+        translate_sentences_api(sentences, args.slang, args.lang, args.output)
+        exit(0)
 
     augmentation_sequence = create_augmentation_sequence(
         args.aug, args.translate_mode, args.lang, args.device
